[WIP][Experimental] RAII-like instrumentation
This introduces a type that based on instantiation and destruction calls
the instrumentation callback with the appropriate `.pre|.post` timing
for the provided info and callback kind.

This cleans up the call sites for instrumentation nicely (no need to
manually wrap pre and post calls around single lines or use `defer`
directly). However, it does introduce an object that gets created and
then never mutated. In Swift these are usually `_` so that they are
ignored. I'm not clear on what the compiler would do: could it completey
elide the function call since it thinks the resulting object is unused?
could it immediately deallocate the returned object since nothing had a
strong reference?

Also the use of closures is not my favorite. I considered something like
```swift
let newLocalState: LocalState; do {
  let inst = instrumentation(callbackInfo, .scopedStoreToLocal)
  newLocalState = toLocalState(newValue)
}
```

But the above feels significantly different than usual Swift and so did
not use it.

Another downside associated with the unused returned object are the
warnings output by the compiler:
```
initialization of immutable value 'inst' was never used; consider replacing with assignment to '_' or removing it
```

Doesn't seem great to have warnings baked into the library. Is there a
better way to model this RAII like behavior?