Instrumentation injected at init and mutable
Instrumentation now supports having the `callback` and
`viewStoreCreated` closures updated at runtime. There is the assumption
that these will be mutated only on the main queue (or at least the same
queue upon which the root `Store` is affined).

Additionally, now the instrumentation instance is injected to the root
`Store` at initialization time and passed along to child stores as
needed.